Add files via upload
https://neurovenge.antonomase.fr/


Evolution of the academic influence of connectionist and symbolic approaches to AI.

The main plot (top) represent the evolution of the ratio between the number of cited publications in the "connectionist" corpus2 (orange) and the corresponding number in the "symbolic" corpus3 (blue), both normalised by the total number of publication indexed by Web of Science. The plots at the bottom represent for each corpus, the number of cited publications cited during the given period.